# attached_assets/Legal-Flow/lib/api-zod/src/generated/types/case.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { CasePriority } from './casePriority';
import type { CaseStatus } from './caseStatus';
import type { Uuid } from './uuid';

export interface Case {
  id: Uuid;
  userId: string;
  clientId?: Uuid | null;
  lawyerId?: string | null;
  assistantId?: string | null;
  caseNumber: string;
  title: string;
  description?: string | null;
  category?: string | null;
  status: CaseStatus;
  priority?: CasePriority;
  createdAt: Date;
  updatedAt: Date;
}