# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LegalFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/sync:
    post:
      operationId: syncCurrentUser
      tags: [users]
      summary: Sync current user from Clerk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSyncInput"
      responses:
        "200":
          description: Synced user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin/owner only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{userId}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update user role (admin/owner only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Clients ────────────────────────────────────────────────────────────────
...
  /cases:
    get:
      operationId: listCases
      tags: [cases]
      summary: List cases
      responses:
        "200":
          description: List of cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Case"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createCase
      tags: [cases]
      summary: Create a case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseInput"
      responses:
        "201":
          description: Created case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cases/{caseId}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: Case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCase
      tags: [cases]
      summary: Update a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseUpdate"
      responses:
        "200":
          description: Updated case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCase
      tags: [cases]
      summary: Delete a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Tasks ──────────────────────────────────────────────────────────────────
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tasks/{taskId}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Documents ──────────────────────────────────────────────────────────────
...
components:
  schemas:
    Uuid:
      type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        firmName:
          type: ["string", "null"]
        initials:
          type: ["string", "null"]
        role:
          type: string
        telegramChatId:
          type: ["string", "null"]
        telegramNotificationsEnabled:
          type: boolean
        onboardingComplete:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - email
        - role
        - telegramNotificationsEnabled
        - onboardingComplete
        - createdAt
        - updatedAt
    UserSyncInput:
      type: object
      properties:
        email:
          type: string
        name:
          type: ["string", "null"]
        firmName:
          type: ["string", "null"]
        initials:
          type: ["string", "null"]
        telegramChatId:
          type: ["string", "null"]
        telegramNotificationsEnabled:
          type: boolean
        onboardingComplete:
          type: boolean
      required:
        - email
    UserRoleUpdate:
      type: object
      properties:
        role:
          type: string
      required:
        - role

    # ── Clients ──
...
    Case:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        clientId:
          type: ["string", "null"]
        lawyerId:
          type: ["string", "null"]
        assistantId:
          type: ["string", "null"]
        caseNumber:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - caseNumber
        - title
        - status
        - createdAt
        - updatedAt
    CaseInput:
      type: object
      properties:
        clientId:
          type: ["string", "null"]
        lawyerId:
          type: ["string", "null"]
        assistantId:
          type: ["string", "null"]
        caseNumber:
          type: ["string", "null"]
        title:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: ["string", "null"]
      required:
        - title
    CaseUpdate:
      type: object
      properties:
        clientId:
          type: ["string", "null"]
        lawyerId:
          type: ["string", "null"]
        assistantId:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: ["string", "null"]

    # ── Tasks ──
    Task:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        clientId:
          type: ["string", "null"]
        caseId:
          type: ["string", "null"]
        assigneeId:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - title
        - status
        - createdAt
        - updatedAt
    TaskInput:
      type: object
      properties:
        clientId:
          type: ["string", "null"]
        caseId:
          type: ["string", "null"]
        assigneeId:
          type: ["string", "null"]
        title:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
      required:
        - title
    TaskUpdate:
      type: object
      properties:
        clientId:
          type: ["string", "null"]
        caseId:
          type: ["string", "null"]
        assigneeId:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]

    # ── Documents ──
...